requirements.txt
streamlit
google-generativeai
python-docx
pdfplumber